Add: FBX file loader. Mov: Effects are now in their own subdirectory and files. Add: gif for int13h splash screen. Add: SSAO effect and shaders. Rem: Some unused files. Add: gBuffer renderer for MRT rendering. Upd: Restyled all the code, hense the HUGE change log. Add: tools.getTangent(). Upd: texture system doesn't bind to next available texture unit, this was silly =P Add: shader var parser supports uniform struct arrays. Fix: shader onLoad runs after newShader is created. Fix: misc bits

+++ shaders/depthmap.frag.shader
@@ -1,6 +1,6 @@
-precision mediump float;
+precision highp float;
 uniform float uType;
 
-void main(void) {
-	gl_FragColor=vec4(gl_FragCoord.z,uType,0.0,1.0);
+void main (void) {
+	gl_FragColor = vec4(gl_FragCoord.z, uType, 0.0, 1.0);
 }
